Cloud Foundry provider
This change adds a Cloud Foundry provider to the dpl utility.  The
implementation of this provider delegates to the cf command line
utility because the internal implementation of the gem does not lend
itself to being called programmatically.

The provides requires five options to be set:

1. target: The API endpoint of the Cloud Foundry instance
2. username: The username to authenticate with
3. password: The password to authenticate with
4. organization: The Cloud Foundry organization to log into
5. space: The Cloud Foundry space to log into

Given this information, the provider targets the specified API
endpoint and logs into the appropriate organization and space.  It
then pushes the application to Cloud Foundry.  Once complete the
provider logs out of Cloud Foundry, removing the authentication
credentials used earlier.