Use specific version of aws-sdk for lambda provider.
This commit makes the lambda provider temporarily depend
on 2.0.37. This is because of the fact that when the
provider was implemented Lambda was in preview (version
2014-11-11). Now the Lambda API is at version 2015-03-31.

The new API is different than the old one, which breaks
the provider for this project.

This is a fix so that the existing provider works, but
the provider needs to be updated to conform to the newest
API.

The patch version has been bumped.