Allow users to specify a package_glob we will search.
In some cases, eligible package or package-like objects may be laying
around (particularly .gem files in .bundle/ from a bundler driven test
suite). This allows the user to specify the glob(s) for the paths we
will search rather than a recursive search of everything.

This adds the option package_glob to the provider. It takes a string
or list of strings. It defaults to "**/*" (search everthing for package
files).