gae: Adapt to gcloud-0.9.80
The new gcloud version deprecates --set-default and introduces --promote
(and --no-promote for transition, apparantly). This is reflected in the
options to the provider. There is no smooth transition from 'default' to
'no_promote' as the provider is experimental.

Also, the release notes hint that --no-stop-previous-version will become
available as soon as --promote kicks in. This is also exposed as an
option.

See https://dl.google.com/dl/cloudsdk/release/RELEASE_NOTES