Handle files with spaces correctly in encoding_for
Fixes https://github.com/travis-ci/travis-ci/issues/5004

When passed a file name with a space, we end up calling
`file` without escaping the spaces (and other special characters).
This most likely causes calling `file` on a nonexistent file.
The upload then appears to hang, and eventually the job is terminated.