Add `#api_endpoint` to Releases provider
This method consults options "host" and "api_version" to
construct the API endpoint that is suitable for both
github.com and GitHub Enterprise.

We will assume that there will be only one API version
on github.com at all times; i.e., "api_version" is
taken into consideration only when "host" is given
(which suggests that GitHub Enterprise is in use).

Note that giving "host: api.github.com" is invalid,
since that does not match the API endpoint expected
by this provider.