Allow Github release provider to overwrite files
Sometimes we want to updating an existing release with a new file that
has the same name. While best practice for user-facing releases is to
make a brand new release, there are some use cases where releases are
used as well known endpoints to grab the latest code.

This commit adds support for an `overwrite` option for the releases
provider, which will upload files even if others with the same name
are present on the release.