Updated the default PyPI upload URL
The "twine" tool uses this URL by default. The only reason why Travis still uses the legacy URL is because it's overridden in ~/.pypirc by this code. This new URL points to the newer Warehouse server which is the replacement for the old Cheeseshop PyPI implementation. Uploading to Warehouse adds the distribution to both servers (unlike the legacy URL which sadly does not seem to affect Warehouse), and has proven more reliable than the older server.