Switch to Python 2.7 in GAE before deploying
Resolves https://github.com/travis-ci/travis-ci/issues/6913

Google Cloud SDK currently requires 2.7, so if the build is
executed with Python 3, deployment fails.

It is possible that another deployment provider in use requires
Python 3 and we may have to switch Python versions back and forth,
but for now, we will assume that we only need to switch to 2.7,
and remain there.