Add option to disable "Deployed at [time]" file
Adds a simple option to be set on `.travis.yml` deploy section, just as the others, to be able to disable the creation of such file. It could be deleted with a complicated after_deploy but I don't see why a simple option cannot be provided.

Update pages.rb

Update pages.rb

Add new option to README