Drop verification of existing firebase-tools installation
This prevents https://github.com/travis-ci/travis-ci/issues/8637 from happening.

The check uses the PATH, which in case of node.js projects includes `node_modules/.bin`. If the firebase-tools package has been installed locally, the check will pass because they are stored in said folder, so the tools will not get installed globally. The problem is that when the deploy is run, `node_modules/.bin` isn't in the PATH so it results in 'sh: 1: firebase: not found' when deploying.