Fix GAE deployable files to accept more than one
Per Google Cloud's `gcloud` documentation[0], the
`app deploy` subcommand accepts multiple types of
deployable files at once. Previously, the string
value was only added to the command string. This
made deploying with multiple files fail as Ruby
fetched the string value of the Array type instead.

Now, the value is only appended to the command
string directly if the value type is a string. If
the value type is an Array, each value is appended
onto the string separately.

`travis.yml` example:

deploy:
 provider: gae
 project: example-1
 keyfile: example-key.json
 config:
  - server/app.yaml
  - server/dos.yaml
  - server/cron.yaml

(or)

deploy:
 provider: gae
 project: example-1
 keyfile: example-key.json
 config: server/app.yaml

--
0: https://cloud.google.com/sdk/gcloud/reference/app/deploy