Tweak github pages deploy process
If `keep_history` was given we have, so far, brute force git cloned
the repo, which may have left a defunct directory, which we have then
fixed by `rm -rf` and `mkdir -p`.

Instead we can check the remote repo to see if it has the given branch
name, and conditionally clone or initialize the repo.

Also, this commit makes the default for `keep_history` true. If users
want to overwrite their history they will need to pass `no_keep_history`
from now on.