Fail CodeDeploy when wait_until_deployed fails
Specifically, if the eventually returned status is not 'Succeeded', then
the deploy fails with a message indicating that waiting was terminated
and the returned status.

This change includes a test that asserts that the failure is correctly
raised in this case.

With some additional effort I could set up a script that should provoke
a non-'Succeeded' status code and prove out the integration itself,
please let me know if that's desired.

I'm definitely also open to feedback on the style of the returned
error message.

Closes: #1186